Split the audio functionality and added repeat
The mechanics of the ALSA audio layer have been separated out into
a new file (with a view to making it more portable). Eventually I
will add SDL2 audio or OSS audio to make the code more portable.
In addition, I added a -r (repeat) option so it will repeat the
string.